POLIO-2125: split form a serializer
