+++ plugins/polio/api/vaccines/stock_management/outgoingstockmovement/serializers/constants.py
@@ -1,0 +1,2 @@
+# Set True by OutgoingStockMovementViewSet.get_serializer_context on write actions.
+VALIDATE_FORM_A_LIFECYCLE_CONTEXT_KEY = "validate_form_a_lifecycle"
